+++ src/lightning/fabric/plugins/precision/amp.py
@@ -41,6 +41,12 @@ class MixedPrecision(Precision):
         device: str,
         scaler: Optional[torch.cuda.amp.GradScaler] = None,
     ) -> None:
+        if precision not in ("16-mixed", "bf16-mixed"):
+            raise ValueError(
+                f"Passed `{type(self).__name__}(precision={precision!r})`."
+                " Precision must be '16-mixed' or 'bf16-mixed'."
+            )
+
         self.precision = cast(Literal["16-mixed", "bf16-mixed"], str(precision))
         if scaler is None and self.precision == "16-mixed":
             with _patch_cuda_is_available():