+++ tests/tests_pytorch/strategies/test_ddp.py
@@ -101,7 +101,7 @@ def test_ddp_wrapper(tmpdir, precision):
     class CustomCallback(Callback):
         def on_train_start(self, trainer: "pl.Trainer", pl_module: "pl.LightningModule") -> None:
             assert isinstance(trainer.strategy.model, DistributedDataParallel)
-            expected = ["module.something"]
+            expected = ["something"]
             assert (
                 trainer.strategy.model.parameters_to_ignore == set(expected) if _TORCH_GREATER_EQUAL_2_0 else expected
             )